requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.8
intasend-python==1.1.2
flask-cors==4.0.0
werkzeug==2.3.7
sqlalchemy==2.0.23